+++ src/lib.rs
@@ -1,10 +1,10 @@
 //! Some useful functions on [`proc_macro`] and [`proc_macro2`] types
 //!
 //! E.g. [pushing tokens onto `TokenStream`](TokenStreamExt::push) and [testing
-//! for specific punctuation on `TokenTree` and Punct](TokenTreePunct)
+//! for specific punctuation on `TokenTree` and `Punct`](TokenTreePunct)
 //!
 //! It also adds the [`assert_tokens!`] and [`assert_expansion!`] macros to
-//! improve unit testability for proc-macros.
+//! improve unit testability for `proc-macros`.
 #![warn(clippy::pedantic, missing_docs)]
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 #![deny(rustdoc::all)]